Moved set_step, submit_code, evaluate to stepikclient
